Section on obtaining steady state probs
I do not like the whole diff-diff equations, it's a bit clumsy and
arguably shows that you do not realise what a MC is (I know this is not
true, it just could be understood that way). The first approach (the
diff-diff thing) is 'using Markov chains' it's just that you're writing
out the equations slowly and going through the basics of what a Markov
chain (without mentioning) it from first principles. It's not terribly
useful.